mychat/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.16.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
isodate==0.6.1
jiter==0.5.0
joblib==1.4.2
jose==1.0.0
jwcrypto==1.5.6
jwt==1.3.1
looseversion==1.3.0
lxml==5.3.0
networkx==3.2.1
nibabel==5.3.1
nipype==1.8.6
nltk==3.9.1
numpy==2.0.2
openai==1.51.0
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pgvector==0.3.5
pluggy==1.5.0
prov==2.0.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
PyJWT==2.9.0
PyMuPDF==1.24.11
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.9.11
requests==2.32.3
scipy==1.13.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
starlette==0.38.6
tomli==2.0.2
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
websockets==13.1
zipp==3.20.2
